Allow newly created state objects to be accessible in other snapshots
Relnote: """
Newly created state objects are immediately accessible from
other snapshots, including the global snapshot, with their initial
state. Any subsequent modifications are not visible until the snapshot
in which the object was created is applied.

The initial state of `mutableStateOf()`, as well as their primitive
versions, is the value passed in as a parameter to `mutableStateOf()`.
The initial state of `mutableStateListOf()` and `mutableStateMapOf()`
is empty.

Added `Snapshot.isInSnapshot` and `PREEXISTING_SNAPSHOT` that are used
to enable a state object to support being immediately accessible from
the global snapshot upon its creation.

Custom state objects can support being immediately accessible by
following the pattern of changes made to the built-in snapshot
objects.
"""

Test: ./gradlew :compose:r:r:tDUT
Fixes: 329243608
Change-Id: I822dbda9c84f430d95444e0234f65b371c41352e