Use PUBLICATION in SupportConnectionPool's Lazy
Calls to useConnection are heavily concurrent and have to be protected a bit. The current impl of UnsafeLazyImpl nulls the initializer in an unprotected way which can lead to NPEs. We are not interested in full locking since the object is just a wrapper so it is OK to return different instances when the underlying object is the same.

Bug: 329701026
Test: n/a
Change-Id: Id8eda96364ce0c3e4c5cd3e20d1a31b29a35dca9