Update Pager semantics to use correct scroll/maxscroll values.
Pager was using the LazyLayout strategy of calculating pseudo scrolls and this was causing the talkback to think that scrolling was finished in some cases, leaving the Pager. In this CL I'm updating how we calculate scroll and maxscroll for pager since these values are available due to Pager's fixed size page nature.

Test: Previous tests should pass.
Relnote: N/A
Fixes: 323414369
Fixes: 321782800
Change-Id: I2294fc38a84580a57c348141912c9dde36a7a5fb