Update SearchBar to use a custom layout
To support inputField becoming a slot, SearchBar needs to
measure the component rather than relying on internal
knowledge of its size.

The Surface that used to wrap the entire layout has been
changed to a child of the Layout to simplify calculations.
This also means we can directly calculate insets/paddings
in the layout lambda rather than relying on modifiers.

Also fixes a predictive back bug for RTL.

Bug: b/326627700
Test: unit tests and screenshots still pass
Change-Id: If036829758665c3c2bda4582caa2ebc4007cd7c6