Fix 'notNull' in Field bundle being optional.
A mistake during migration to kotlinx-serialization incorrectly made it not null, but a long time ago Room didn't emit schema files with 'notNull' and thus to keep reading those old files it must be and optional entry.

Also went back in the commit history and added tests for other optionals. To help catch this I ended up making the 'Unable to read schema file' warning an error, since once there is a processor errors, warnings are not displayed. Its more noisy but waaay more helpful.

Lastly, noticed default indentation with kotlinx-serialization is 4 spaces, made it 2 as that is what we used.

Bug: 329865354
Test: SerializationTest
Change-Id: I81be22e95a146148b5943ee145fb27c69d186c74