OutOfBoundsCount changes were causing extra recompositions.
In the latest changes out of bonds param count were being considered to trigger recomposition in the new logic for saving recomp for visible pages changes. The logic should actually be based on visible pages only, since if the visible changes don't change, there's no reason for the out of bounds pages to change. We must, however, apply the delta changes to the invisible pages as well.

Test: Updated existing tests
Fixes: 329702052
Change-Id: Ie3317dc8aefb9289d69991c4d6b815c36d5e0771