NavGraphBuilder adds support for Object as startDestination for root NavGraph
Add a secondary constrcutor that takes a serializable Object as startDestination, as well as DSL methods to create a NavGraph / nested NavGraph with object as startDestinations.

Test: ./gradlew navigation:navigation-common:cC
Bug: 188693139
Relnote: "NavGraphBuilder can now take a serializable Object as startDestination. The starting NavDestination must have been added as a KClass in order to match."
Change-Id: I20a147e037ff80f0cab34276275890e683fa8367