Use .values instead of .entries for enum converters
The .entries API was only introduced in Kotlin 1.9.0, so does not work
in older versions. The reason for .values() getting (soft) deprecated
is that it causes a new allocation each time it is called. We work
around this by creating a single copy in the converter which we reuse.

See also: https://github.com/Kotlin/KEEP/blob/master/proposals/enum-entries.md

Test: ./gradlew :privacysandbox:tools:tools-{apigenerator,apicompiler,apipackager,core}:test
Change-Id: I7ad0890f26781c253f2d3d972dc8c4b85eb23693