Implement NavController popBackStack for Safe Args
Added popBackStack overloads to pop with a serializable KClass or with a serializable Object.

For KClass version, there are no arguments values to consider so we can pop directly based on its id. For Object version, we need to generate the route with args filled and then pop with that.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 188693139
Relnote: "Added NavController popBackStack APIs for popping from a serializable KClass or from a serializable Object. The target NavDestination to be popped must have been created with route from KClass."
Change-Id: I1033daaefa341eb1054a9b00b0fce0b5317adca8