Implement NavController clearBackStack for Safe Args
Added clearBackStack overloads to clear with a serializable KClass or with a serializable Object.

For KClass version, there are no arguments values to consider so we can clear directly based on its id. For Object version, we need to generate the route with args filled and then clear with that.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 188693139
Relnote: "Added NavController clearBackStack APIs for clearing saved state with a serializable KClass or from a serializable Object. The target NavDestination to be cleared must have been created with route from KClass."
Change-Id: I8972f003efa07fd275d397bf31237d6b3f3e38d1