Do not reset animators when lazy lists are scrolled to the same position
Calling state.scrollToItem() with passing the same position as it is now is used in order to disable the scroll position maintaining by key and is not considered to be a real scroll by the users. In this case we need to still apply all the animations if also the data set is changing. Before this change we would just reset the animation state completely disabling all the animations.

Test: LazyListAnimateItemPlacementTest, LazyGridAnimateItemPlacementTest, LazyStaggeredGridAnimateItemPlacementTest
Change-Id: I2fbeede2362f7ff9440453ede25764e8cceab60c