Implement NavController navigate with Object T
Add support for navigating with a serializable object instance. The target NavDestination must have been added to the NavGraph with a route from a KClass.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 188693139
Relnote: "Add new navigate API to navigate with a serializable object instance. The target NavDestination must have been added to the NavGraph with a route from KClass."
Change-Id: If44d34757c454d628e2e73794047a5a23c37308a