Change the default contentAlignment to TopStart
Moving the NavHost default contentAlignment to `Alignment.TopStart` in
order to match the defaults of AnimatedContent. We previously used
`Alignment.Center` to match the previous used `CrossFade` API, but it
would cause unexpected position transitions when using specific
component such as column, row, etc.

RelNote: "`NavHost` now used `Alignment.TopStart` as the default
contentAlignment argument. This puts it in line with the default for
`AnimatedContent` and fixes some instances of an unexpected scale from
center transition."
Test: tested in sample app
Bug: 330111602

Change-Id: I09e7260d31cb571039629acd83b58eaf000a2493