Add link styling arguments to the parseAsHtml
Styling properties like color set directly on the html string for links will be merged with (and overwritten in case of the same key) by
those coming from the arguments in this method. For example, in string like "<a href>link1</a><span style="color:red"><a href>link2</a></span> is you construct an AnnotatedString with parseAsHtml(linkStyle = SpanStyle(color = green)), both link1 and link2 will be rendered as green. Eventually we could add an argument 'overrideLinkStyle' boolean or similar that would allow developers to preserve the styling set directly in the html, in this case that would allow to render the second link as red.

Relnote: Added styling arguments (linkStyle, focusedLinkStyle, hoveredLinkStyle) and a link interaction listener to the `parseAsHtml` method. When parsing the HTML-tagged string with <a> tags, the method will construct a LinkAnnotation.Url for each such tag and pass the styling objects and link interaction listener to each annotation.
Test: AnnotatedStringFromHtmlTest

Change-Id: I7c9777a340e04ccf4dc10258c83d18e69831b3c6