Fix transaction with non-seekable sharedElement
If any transition that is part of the transaction is not seekable, then
the entire transaction should not be seekable. That includes shared
element transitions. There is currently a case where if there is a
enter/exit transition that is seekable along with a
sharedElementEnter/Return that is not seekable, fragments will cause the
Transition library to crash by sending the sharedElement transition down
the seeking path.

We currently don't consider shared element transitions seekability at
all, allowing us to ensure we don't attempt to seek a shared element if
no other transition exists. But we also should ensure that if there is a
shared element present along with other seekable transitions, that
shared element transition is seekable as well.

RelNote: "Fixed a bug where if a non-seekable shared element was added
to a transactions where all other transitions were seekable, there would
be a crash. Now the transaction will correctly be considered
non-seekable."
Test: added FragmentTransitionSeekingTest
Bug: 330852775
Bug: 300657793

Change-Id: I18ccdbc4f7226f265106435945f8ee4039dc475b