Fixed removing movable content followed by a remember
If movable content is removed in the same frame as a key'ed
remember was updated it could result in an index out of bounds
exception.

This was caused by remember obtaining an anchor to a group
in the movable content that was removed. If the code was simply
removed then anchor was correctly checked for validity and
ignored. However, since the content is still retained (until
the end of the frame) its anchor is still valid and would
then be used in the wrong table.

The fix is to ensure that the remember holder only takes an
anchor for a sibling group, not a group that will be moved
into a movable content holder.

Fixes: 305971779
Test: ./gradlew :compose:r:r:tDUT
Change-Id: I306f8e3caaf90644d0059ecf5e1b6e34c75b2d1c