Make TextFieldCharSequence internal
TextFieldState and TextFieldBuffer now expose the properties of TextFieldCharSequence as separate entities on their surface. This enables us to optimize the implementation details in the future by not committing to a single class on the API surface.

Test: existing tests
Relnote: "Removed `TextFieldCharSequence`. `TextFieldBuffer.originalValues` is replaced with `TextFieldBuffer.originalText` and `TextFieldBuffer.originalSelection`."
Change-Id: I2c7d6d00ddd9a6412ed97f4ec893f97b0e2342c9