Introduce AnimatedPaneScope
It doesn't make sense to pass ThreePaneScaffoldScope to the content of
AnimatedPane as it's already avaiable to devs from the pane composables.
Introduce AnimatedPaneScope that extends AnimatedVisibilityScope to
provide more info to the actual pane content. It can also be expanded
in the future to provide more relevant info.

Relnote: introduce AnimatedPaneScope
Bug: 332750742
Test: n/a
Change-Id: I62d73a2322ad61b1734bfdb9782e5644741887f4